+++ noj-core/src/app.ts
@@ -6,9 +6,9 @@ import stats from "./routes/stats.ts";
 import auth from "./domains/identity/routes/auth.ts";
 import admin from "./routes/admin/index.ts";
 import adminAnnouncements from "./routes/admin-announcements.ts";
-import adminTrainings from "./routes/admin-trainings.ts";
-import tags from "./routes/tags.ts";
-import problems from "./routes/problems.ts";
+import adminTrainings from "./domains/catalog/routes/admin-trainings.ts";
+import tags from "./domains/catalog/routes/tags.ts";
+import problems from "./domains/catalog/routes/problems.ts";
 import checkin from "./domains/identity/routes/checkin.ts";
 import queue from "./routes/queue.ts";
 import submissions from "./routes/submissions.ts";
@@ -20,7 +20,7 @@ import community from "./routes/community.ts";
 import communityAdmin from "./routes/community-admin.ts";
 import search from "./routes/search.ts";
 import contests from "./domains/contest/routes/contests.ts";
-import trainings from "./routes/trainings.ts";
+import trainings from "./domains/catalog/routes/trainings.ts";
 import announcements from "./routes/announcements.ts";
 import sse, { contestSse, statsSse } from "./routes/sse.ts";
 import { AppError } from "./lib/errors.ts";
