+++ noj-core/src/domains/catalog/index.ts
@@ -1,0 +1,11 @@
+export * from "./services/problems/problems.ts";
+export {
+  applyAlgorithmTagVisibility,
+  attachTags,
+  type ProblemViewer,
+} from "./services/problems/problems-list.ts";
+export * from "./services/problems/problem-bundle.ts";
+export * from "./services/problems/problem-field-guard.ts";
+export * from "./services/tags.ts";
+export * from "./services/trainings.ts";
+export * from "./services/support-package.ts";
