+++ noj-core/src/domains/catalog/services/problems/problems-tags.ts
@@ -13,9 +13,9 @@
  * 使校验失败（400）不产生孤儿题目或已提交字段（syncProblemTags 内部仍重复校验兜底）。
  */
 import { eq, inArray } from "drizzle-orm";
-import { getDb } from "../../db/connection.ts";
-import { problemTags, tags } from "../../db/schema.ts";
-import { BadRequestError } from "../../lib/errors.ts";
+import { getDb } from "../../../../db/connection.ts";
+import { problemTags, tags } from "../../../../db/schema.ts";
+import { BadRequestError } from "../../../../lib/errors.ts";
 
 /**
  * 写库前校验标签集合（去重后）：
