+++ noj-core/src/domains/gateway/index.ts
@@ -1,6 +1,7 @@
 export {
   createUserLlmProvider,
   deleteUserLlmProvider,
+  getLlmProviderById,
   listUserLlmProviders,
   LlmGatewayError,
   testUserLlmProvider,
